Show first thread author avatar in forum channels
Forum channels now display the avatar of the first thread's author instead of the server icon. This is achieved by extracting the first thread author in the loader and passing it to the content component, which conditionally renders the avatar. Also refactored the SuspenseClientOnly component in code.tsx to use useEffect for client-only rendering.